[NodeJS][Challenge06] - Developed all features following the test files (list, create, import, delete)

+++ src/database/migrations/1588100422510-CreateTransactions.ts
@@ -23,7 +23,7 @@ export class CreateTransactions1588100422510 implements MigrationInterface {
           },
           {
             name: 'value',
-            type: 'integer',
+            type: 'decimal',
           },
           {
             name: 'category_id',
